Add support for writing to multiple graphite hosts with matching rules
The rules are left to the admin to set up.
Default setup is set to 'false' but shows how to send metrics to
three graphite relays with sending only 1/3 of metrics per destination
This way each server gets 2/3 of the traffic.
Can be expanded to more hosts or different rules.

More complex rules were not tested, see example/ directory for more.